renpy.eclass: add RENPY_KEEP_COMPILED

+++ eclass/renpy.eclass
@@ -24,6 +24,12 @@ esac
 # The title of the VN to be used in the desktop entry. Defaults to ${PN}.
 : "${RENPY_TITLE:="${PN}"}"
 
+# @ECLASS_VARIABLE: RENPY_KEEP_COMPILED
+# @DEFAULT_UNSET
+# @DESCRIPTION:
+# If set, the installation will keep rpyc and rpyb files in place instead of
+# deleting them.
+
 # @ECLASS_VARIABLE: RENPY_WINDOW_ICON
 # @DESCRIPTION:
 # Path to the menu icon for installation.
@@ -69,8 +75,10 @@ renpy_src_prepare() {
 	done
 
 	find game -name "*.rpa" -delete || die
-	find game -name "*.rpyc" -delete || die
-	find game -name "*.rpyb" -delete || die
+	if [[ -z "${RENPY_KEEP_COMPILED}" ]]; then
+		find game -name "*.rpyc" -delete || die
+		find game -name "*.rpyb" -delete || die
+	fi
 
 	default
 }
